Imported "aoppa-0.5.4.tar.gz".
Committed by Mikko Juola (and Mikko is writing this commit message); the
author of the code in this repository is authored by Oskari Saarenmaa
and licensed under GPL3. I manually assigned the author to Oskari as
they own copyright over the code and should be noted so :)

I made this commit on 25 May 2025 (but manually set date to what I think
is when the code was actually released, but I had to rely on looking
inside the files).

If there are other authors, it is not apparent with a quick skim on this
codebase.

The tarball I used to initialize this repository was from here:
https://auno.org/dev/aoppa/ and I fetched it 25 May 2025. Last updates I
see are from 2009.

sha256sum of aoppa-0.5.4.tar.gz:
9348d87c2971b4f145582944cefe961d857c2fc8914886853005094a6b23a6d2

Code archeology purposes. This did not compile out-of-box.